move user_profile.py to structures folder
